Added Enemy Follow Script
It's a bit janky currently but the enemy will now follow the player around.
BE ADVISED: As I configured a lot of this in my scene and didn't make it some damn prefabs it won't work until you link up shit properly. So I will do so now and it shouldn't be a problem.

Also, we need to move the movement/input/etc off the player_model and onto the base of the player :)  The model should only be the model and the parent Player object should drive everything