Added Camera Follow script
You can now just drag the Camera Prefab onto your scene anywhere and it should work.
If you want to adjust the position of the camera, just use the Variables in the Camera Follow script instead of physically moving the camera with its Transform component.
BE SURE TO REFERENCE THE PLAYER in the Target variable in the Inspector.