Organised Assets, Polished Player Healthbar, created enemy healthbar, finished rapier animation, created damage function to enemies by rapier.
Created folders for assets.
Player Healthbar is in top lef tof screen and takes damage oinb collision with enemies.
Enemy Healthbar currently on bottom left of screen and takes damage on collision with the rapier. this is based on timings of the animation turning box collider(trigger) on and off.