Added death animation to enemy. Cleaned up Enemy script.
Enemies now play death animation and disable their colliders and navmesh components when they die. Current problem is the Translation of the armature is above ground so they float, and something causes them to spin on the spot.

Removed old code, we can check GitHubs history if we need to go back but there is currently no point having it in there.